refactor: clean up the clean code
