+++ src/clean.ts
@@ -4,43 +4,27 @@ import MongoObject from 'mongo-object'
 import convertToProperType from './clean/convertToProperType.js'
 import setAutoValues from './clean/setAutoValues.js'
 import { SimpleSchema } from './SimpleSchema.js'
-import { CleanOptions } from './types.js'
+import { CleanOptions, NodeContext } from './types.js'
 import { isEmptyObject, looksLikeModifier } from './utility/index.js'
-import typeValidator from './validation/typeValidator/index.js'
-
-interface ForEachNodeContext {
-  genericKey: string
-  isArrayItem: boolean
-  operator: string
-  position: string
-  remove: () => void
-  updateValue: (newValue: any) => void
-  value: any
-}
+import { isValueTypeValid } from './validation/typeValidator/index.js'
 
 const operatorsToIgnoreValue = ['$unset', '$currentDate']
 
+function log (message: string): void {
+  if (SimpleSchema.debug === true) {
+    console.debug(message)
+  }
+}
+
 /**
- * @param {SimpleSchema} ss - A SimpleSchema instance
- * @param {Object} doc - Document or modifier to clean. Referenced object will be modified in place.
- * @param {Object} [options]
- * @param {Boolean} [options.mutate=false] - Mutate doc. Set this to true to improve performance if you don't mind mutating the object you're cleaning.
- * @param {Boolean} [options.filter=true] - Do filtering?
- * @param {Boolean} [options.autoConvert=true] - Do automatic type converting?
- * @param {Boolean} [options.removeEmptyStrings=true] - Remove keys in normal object or $set where the value is an empty string?
- * @param {Boolean} [options.removeNullsFromArrays=false] - Remove all null items from all arrays
- * @param {Boolean} [options.trimStrings=true] - Trim string values?
- * @param {Boolean} [options.getAutoValues=true] - Inject automatic and default values?
- * @param {Boolean} [options.isModifier=false] - Is doc a modifier object?
- * @param {Boolean} [options.isUpsert=false] - Will the modifier object be used to do an upsert? This is used
- *   to determine whether $setOnInsert should be added to it for defaultValues.
- * @param {Boolean} [options.mongoObject] - If you already have the mongoObject instance, pass it to improve performance
- * @param {Object} [options.extendAutoValueContext] - This object will be added to the `this` context of autoValue functions.
- * @returns {Object} The modified doc.
- *
  * Cleans a document or modifier object. By default, will filter, automatically
  * type convert where possible, and inject automatic/default values. Use the options
  * to skip one or more of these.
+ *
+ * @param ss A SimpleSchema instance
+ * @param doc Document or modifier to clean. Referenced object will be modified in place.
+ * @param options Clean options
+ * @returns The modified doc.
  */
 function clean (
   ss: SimpleSchema,
@@ -72,7 +56,7 @@ function clean (
     const removedPositions: string[] = [] // For removing now-empty objects after
 
     mongoObject.forEachNode(
-      function eachNode (this: ForEachNodeContext) {
+      function eachNode (this: NodeContext) {
         // The value of a $unset is irrelevant, so no point in cleaning it.
         // Also we do not care if fields not in the schema are unset.
         // Other operators also have values that we wouldn't want to clean.
@@ -99,11 +83,7 @@ function clean (
             this.remove()
             removedPositions.push(this.position)
           }
-          if (SimpleSchema.debug === true) {
-            console.info(
-              `SimpleSchema.clean: filtered out value that would have affected key "${gKey}", which is not allowed by the schema`
-            )
-          }
+          log(`SimpleSchema.clean: filtered out value that would have affected key "${gKey}", which is not allowed by the schema`)
           return // no reason to do more
         }
 
@@ -112,63 +92,47 @@ function clean (
         const def = defs?.[0]
 
         // Auto-convert values if requested and if possible
-        if (cleanOptions.autoConvert === true && def != null) {
-          const isValidType = defs.some((definition) => {
-            const typeValidationError = typeValidator.call({
-              valueShouldBeChecked: true,
-              definition,
-              value: val,
-              operator: this.operator
-            })
-            return typeValidationError === undefined
-          })
-
-          if (!isValidType) {
-            const newVal = convertToProperType(val, def.type)
-            if (newVal !== undefined && newVal !== val) {
-              if (SimpleSchema.debug === true) {
-                console.info(
-                  `SimpleSchema.clean: auto-converted value ${String(val)} from ${typeof val} to ${typeof newVal} for ${gKey}`
-                )
-              }
-              val = newVal
-              this.updateValue(newVal)
-            }
+        if (cleanOptions.autoConvert === true && defs !== undefined && def != null && !isValueTypeValid(defs, val, this.operator)) {
+          const newVal = convertToProperType(val, def.type)
+          if (newVal !== undefined && newVal !== val) {
+            log(`SimpleSchema.clean: auto-converted value ${String(val)} from ${typeof val} to ${typeof newVal} for ${gKey}`)
+            val = newVal
+            this.updateValue(newVal)
           }
         }
 
-        // Trim strings if
-        // 1. The trimStrings option is `true` AND
-        // 2. The field is not in the schema OR is in the schema with `trim` !== `false` AND
-        // 3. The value is a string.
-        if (
-          cleanOptions.trimStrings === true &&
-          def?.trim !== false &&
-          typeof val === 'string'
-        ) {
-          val = val.trim()
-          this.updateValue(val)
-        }
+        // Clean string values
+        if (typeof val === 'string') {
+          // Trim strings if
+          // 1. The trimStrings option is `true` AND
+          // 2. The field is not in the schema OR is in the schema with `trim` !== `false`
+          if (
+            cleanOptions.trimStrings === true &&
+            def?.trim !== false
+          ) {
+            val = val.trim()
+            this.updateValue(val)
+          }
 
-        // Remove empty strings if
-        // 1. The removeEmptyStrings option is `true` AND
-        // 2. The value is in a normal object or in the $set part of a modifier
-        // 3. The value is an empty string.
-        if (
-          cleanOptions.removeEmptyStrings === true &&
-          (this.operator == null || this.operator === '$set') &&
-          typeof val === 'string' &&
-          val.length === 0
-        ) {
-          // For a document, we remove any fields that are being set to an empty string
-          this.remove()
-          // For a modifier, we $unset any fields that are being set to an empty string.
-          // But only if we're not already within an entire object that is being set.
-          if (this.operator === '$set') {
-            const matches = this.position.match(/\[/g)
-            if (matches !== null && matches.length < 2) {
-              p = this.position.replace('$set', '$unset')
-              mongoObject.setValueForPosition(p, '')
+          // Remove empty strings if
+          // 1. The removeEmptyStrings option is `true` AND
+          // 2. The value is in a normal object or in the $set part of a modifier
+          // 3. The value is an empty string.
+          if (
+            cleanOptions.removeEmptyStrings === true &&
+            (this.operator == null || this.operator === '$set') &&
+            val.length === 0
+          ) {
+            // For a document, we remove any fields that are being set to an empty string
+            this.remove()
+            // For a modifier, we $unset any fields that are being set to an empty string.
+            // But only if we're not already within an entire object that is being set.
+            if (this.operator === '$set') {
+              const matches = this.position.match(/\[/g)
+              if (matches !== null && matches.length < 2) {
+                p = this.position.replace('$set', '$unset')
+                mongoObject.setValueForPosition(p, '')
+              }
             }
           }
         }
@@ -190,7 +154,7 @@ function clean (
   }
 
   // Set automatic values
-  cleanOptions.getAutoValues === true &&
+  if (cleanOptions.getAutoValues === true) {
     setAutoValues(
       ss.autoValueFunctions(),
       mongoObject,
@@ -198,6 +162,7 @@ function clean (
       cleanOptions.isUpsert || false,
       cleanOptions.extendAutoValueContext
     )
+  }
 
   // Ensure we don't have any operators set to an empty object
   // since MongoDB 2.6+ will throw errors.
